requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.15.0
pillow==10.2.0
numpy==1.26.4
gunicorn==21.2.0
gdown
